requirements.txt
Django==1.4.1
MySQL-python==1.2.4
South==0.7.6
argparse==1.2.1
dj-database-url==0.2.1
django-taggit==0.10a1
django-tastypie==0.9.14
ipdb==0.7
ipython==0.13.1
lxml==3.1.1
mimeparse==0.1.3
oauthlib==0.4.0
psycopg2==2.4.6
python-dateutil==2.1
requests==1.1.0
requests-oauthlib==0.3.0
six==1.3.0
smartfile==2.1-1
wsgiref==0.1.2
